+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Services/RoleAssignments/RoleAssignmentsMapper.cs
@@ -136,12 +136,12 @@ internal class RoleAssignmentsMapper : IRoleAssignmentsMapper
             {
                 var existingRoleAssignments = roleAssignments
                     .Where(e => e.EntityId == entityId && e.KeyRole == role.Key.Name && e.KeyModule == role.Key.ModuleId)
-                    .Select(x => new GatewayModel.ComponentEntityRoleAssignmentEntryAssignment(GatewayModel.RoleAssignmentResolution.Explicit, new JRaw(x.RoleAssignments)))
+                    .Select(x => new GatewayModel.ComponentEntityRoleAssignmentEntryExplicitAssignment(new JRaw(x.RoleAssignments)))
                     .FirstOrDefault();
 
                 return new GatewayModel.ComponentEntityRoleAssignmentEntry(
                     new GatewayModel.RoleKey(role.Key.Name, role.Key.ModuleId.ToGatewayModel()),
-                    existingRoleAssignments ?? new GatewayModel.ComponentEntityRoleAssignmentEntryAssignment(GatewayModel.RoleAssignmentResolution.Owner, null),
+                    (GatewayModel.ComponentEntityRoleAssignmentEntryAssignment?)existingRoleAssignments ?? new GatewayModel.ComponentEntityRoleAssignmentEntryOwnerAssignment(),
                     role.Updaters.Select(x => new GatewayModel.RoleKey(x.Name, x.ModuleId.ToGatewayModel())).ToList()
                 );
             })